Adding support for non-string federation keys
Summary:
The recent change to support compound keys for federation also limited the key types to strings (including IDs and other types that map to strings): https://phabricator.khanacademy.org/D59569 Updating the generated code to support any type of key.

While testing different key types, I discovered an underlying bug in gqlgen that causes federation to fail for keys that don't come in as existing Go types. For example, a graphql.Int field from Python that is passed in by the gateway as json.Number failed because service.go tried to access it as `int` without unmarshalling it. Adding that unmarshalling call in this diff to support other types.

Test Plan:
- ran `make gqlgen && make serve` against existing content-library service with compound keys (both strings) and queried with GQL
- ran `make gqlgen && make serve` against assignments service which has an ID key, queried with GQL
- changed one key in content-library to an int and ran `make gqlgen && make serve` with GQL queries before and after this fix

Reviewers: dhruv, csilvers

Differential Revision: https://phabricator.khanacademy.org/D59906